deps: add hacky way to check deps
This is to allow travis to build everything, without failing,
and also to ensure 'sane' installs of pygame, and pygtk. The
reason I'm specifying 'sane' is because it seems that pygtk
is built differently on different systems (eg: distulis only
support windows apparently).

Pygame seems to be distro dependant as well. Regardless if
the given distro installs both of the above, the library
should work without problem

Fixes #11